Try to refetch question if it does not contain audio_src_url initially.
This is needed because new questions will never arrive with audio files.

Resemble does not allow synchronous API for new accounts.

We need to wait until their webhook comes through and updates question.

The good thing is that it provides a better UX :)

We can see the answer as soon as it is generated.

And hopefully audio file comes through before the answer is being typed.