Fix truncating algorithm
We look only at the space left

if the separator won't fit we don't bother.

If the fragment wont fit we try to truncate it.

If we truncate too much (blank string) we don't bother.

Finally if it fits we add it.

loop ends if we run out of fragments or space left.