Getrid of numpy and pandas
Apparently we don't need them for the project of this scale.

Which is fortunate because mixing rubygems and python libraries is not the best developer experience.

CSV loading and calculationas in memory work fine.

Especially with new embeddings model which generate smaller files.

If performance matter I'd rather use vector database or pg_vector extension for postgresql.